requirements.txt
python-telegram-bot==21.7
python-dotenv==1.0.1
openai>=1.40.0
